Strange magic bug fix
Without this fix, TaskList coloring dueDate is strangely non deterministic on a working version (to be commited once it is bug-free, this is not the only one).
In the working version, which only changes allowing dueDates to be in the past, some tasks marked as done (any of them, but only a small number) will have colors. When going to TaskFragment, it resets to black, when marking as done, it can go to ANY color (multiple markDone changes color multiple times), when scrolled, it resets to a bad color
This only happends in some