Updated ModalBottomSheet
Added icon to go back from Edit to view.

Refactored the swapping logic between edit/view.
Now only 1 of the Dialogs remains open. This means the view modal is dismissed/popped from stack when going to edit, so now edit can only return to view when using the button to cancel edit.
Some onPause/onResume logic has changed because of this and because the dialog appears to pop the stack after onPause instead of before (relevant for Form -> List transition to allow storing unfinished add data)